remove module user for fix upercase name"
